Move todo to contaienr and form submit straight to firebase.
